#67 Split quizResolver into graphql component and DI class
