+++ src/service.locator.ts
@@ -2,6 +2,7 @@ import { AuthenticationService } from './auth/authentication.service';
 import { AuthorisationService } from './auth/authorisation.service';
 import { S3FileService } from './file/s3.service';
 import { SQSQueueService } from './queue/sqs.service';
+import { QuizService } from './quiz/quiz.service';
 
 // auth
 export const authenticationService = new AuthenticationService();
@@ -12,3 +13,6 @@ export const fileService = new S3FileService();
 
 // queue
 export const queueService = new SQSQueueService();
+
+// quiz
+export const quizService = new QuizService();
